modify kill test and mv it to issue
